+++ electron/src/index.ts
@@ -341,9 +341,10 @@ export class CapacitorSQLite implements CapacitorSQLitePlugin {
       'transaction',
       true,
     );
-
     const readonly: boolean = options.readonly ? options.readonly : false;
-    const returnMode: string = options.returnMode ? options.returnMode : 'no';
+    const returnMode: string = Object.keys(options).includes('returnMode')
+    ? options.returnMode 
+    : 'no';
     const isSQL92: boolean = Object.keys(options).includes('isSQL92')
       ? options.isSQL92
       : true;
@@ -402,7 +403,7 @@ export class CapacitorSQLite implements CapacitorSQLitePlugin {
       ? options.isSQL92
       : true;
 
-    const connName = 'RW_' + dbName;
+      const connName = 'RW_' + dbName;
 
     const database = this.getDatabaseConnectionOrThrowError(connName);