+++ ios/Plugin/Database.swift
@@ -688,6 +688,8 @@ class Database {
             if isExists {
                 // Set the last exported date
                 try ExportToJson.setLastExportDate(mDB: self, sTime: syncTime)
+            } else {
+                throw DatabaseError.exportToJson(message: "No sync_table available")
             }
             // Launch the export process
             let data: [String: Any] = [