+++ ios/Plugin/Utils/UtilsSQLStatement.swift
@@ -323,28 +323,114 @@ class UtilsSQLStatement {
         return lines.joined(separator: " ")
     }
 
+    // MARK: - isReturning
+
+    class func isReturning(sqlStmt: String) -> (Bool, String, String) {
+        var stmt = sqlStmt.replacingOccurrences(of: "\n", with: "")
+                          .trimmingCharacters(in: .whitespacesAndNewlines)
+        if stmt.hasSuffix(";") {
+            // Remove the suffix
+            stmt = String(stmt.dropLast())
+                        .trimmingCharacters(in: .whitespacesAndNewlines)
+        }
+
+        switch (stmt.prefix(6).uppercased())  {
+            
+          case "INSERT":
+            if let valuesIndex = stmt.range(of: "VALUES", options: .caseInsensitive)?.lowerBound,
+               let closingParenthesisIndex = stmt.range(of: ")", options: .backwards, range: valuesIndex..<stmt.endIndex)?.upperBound {
+                    guard closingParenthesisIndex < stmt.endIndex else {
+                        stmt += ";"
+                        return (false, stmt, "")
+                    }
+
+                    let intParenthesisValue = stmt.distance(from: stmt.startIndex, to: closingParenthesisIndex)
+                    let substringAfterValues = stmt[closingParenthesisIndex...]
+                    var resultString = String(substringAfterValues)
+                                .trimmingCharacters(in: .whitespacesAndNewlines)
+                    if resultString.count > 0 && !resultString.hasSuffix(";") {
+                        resultString += ";"
+                    }
+
+                    let substringStartToEndParenthesis = stmt[...closingParenthesisIndex]
+                    let stmtString = String(substringStartToEndParenthesis)
+                                .trimmingCharacters(in: .whitespacesAndNewlines)
+                                .appending(";")
+
+                    if substringAfterValues.lowercased().contains("returning") {
+                        return (true,stmtString,resultString)
+                    } else {
+                        return (false, stmt, "")
+                    }
+                }
+                return (false, stmt, "")
+            
+            
+          case "DELETE", "UPDATE":
+            let words = stmt.components(separatedBy: .whitespacesAndNewlines)
+            var wordsBeforeReturning: [String] = []
+            var returningString: [String] = []
+
+
+            var isReturningOutsideMessage = false
+            for word in words {
+                if word.lowercased() == "returning" {
+                    isReturningOutsideMessage = true
+                    // Include "RETURNING" and the words after it in returningString
+                    returningString.append(contentsOf: [word] + UtilsSQLStatement.wordsAfter(word, in: words))
+                    break
+                }
+                wordsBeforeReturning.append(word)
+            }
+
+            if isReturningOutsideMessage {
+                let joinedWords = wordsBeforeReturning.joined(separator: " ") + ";"
+                var joinedReturningString = returningString.joined(separator: " ")
+                if joinedReturningString.count > 0 &&
+                            !joinedReturningString.hasSuffix(";") {
+                    joinedReturningString += ";"
+                }
+
+                return (true, joinedWords, joinedReturningString)
+            } else {
+                return (false, stmt, "")
+            }
+            
+          default:
+            return (false, stmt, "")
+        }
+        
+    }
+
+    // MARK: - wordsAfter
+
+    class func wordsAfter(_ word: String, in words: [String]) -> [String] {
+        guard let index = words.firstIndex(of: word) else {
+            return []
+        }
+        return Array(words.suffix(from: index + 1))
+    }
+    
     // MARK: - getStmtAndRetColNames
 
     class func getStmtAndRetColNames(sqlStmt: String, retMode: String)
     -> [String: String] {
-        let retWord = "RETURNING"
-
         var retStmtNames: [String: String] = [:]
-        retStmtNames["stmt"] = sqlStmt
+        
+        let (isReturning, stmt, suffix) = isReturning(sqlStmt: sqlStmt)
+        retStmtNames["stmt"] = stmt
         retStmtNames["names"] = ""
-        if let range = sqlStmt.uppercased().range(of: retWord) {
-            let prefix = sqlStmt.prefix(upTo: range.lowerBound)
-            retStmtNames["stmt"] = "\(prefix);"
-            retStmtNames["names"] = ""
-            if retMode.prefix(2) == "wA" {
-                let suffix = sqlStmt.suffix(from: range.upperBound)
+        if isReturning && retMode.prefix(2) == "wA" {
+            let lowercaseSuffix = suffix.lowercased()
+            if let returningIndex = lowercaseSuffix.range(of: "returning") {
+                let substring = suffix[returningIndex.upperBound...]
+
                 let names =
-                    "\(suffix)".trimmingLeadingAndTrailingSpaces()
+                "\(substring)".trimmingLeadingAndTrailingSpaces()
                 if names.suffix(1) == ";" {
                     retStmtNames["names"] = String(names.dropLast())
                 }
             }
-
         }
         return retStmtNames
     }